# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CurrentUserResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/dto.UserArticleResponse'
        type: array
      bio:
        type: string
      email:
        type: string
      followersCount:
        type: integer
      followingCount:
        type: integer
      id:
        type: integer
      image:
        type: string
      username:
        type: string
    type: object
...
  dto.FollowUserRequest:
    properties:
      follow:
        type: boolean
    required:
    - follow
    type: object
  dto.LoginUserRequest:
    properties:
      password:
        maxLength: 100
        type: string
      username:
        maxLength: 255
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginUserResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RegisterUserRequest:
    properties:
      email:
        maxLength: 255
        type: string
      password:
        maxLength: 100
        type: string
      username:
        maxLength: 50
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.RegisterUserResponse:
    properties:
      bio:
        type: string
      email:
        type: string
      id:
        type: integer
      image:
        type: string
      username:
        type: string
    type: object
...
  dto.UserArticleResponse:
    properties:
      article_id:
        type: integer
      title:
        type: string
    type: object
  dto.UserResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/dto.UserArticleResponse'
        type: array
      bio:
        type: string
      email:
        type: string
      followersCount:
        type: integer
      following:
        type: boolean
      followingCount:
        type: integer
      id:
        type: integer
      image:
        type: string
      username:
        type: string
    type: object
  utils.APIErrorResponse:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
...
  utils.ValidationError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
      errors:
        items:
          $ref: '#/definitions/utils.ValidationFieldError'
        type: array
    type: object
  utils.ValidationFieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: REST API for the Realworld application.
  title: Realworld API
  version: "1.0"
paths:
...
  /profiles/{username}:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a user profile
      tags:
      - profiles
  /profiles/{username}/follow:
    post:
      consumes:
      - application/json
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        type: string
      - description: Follow payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.FollowUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Follow or unfollow a user
      tags:
      - profiles
...
  /user:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CurrentUserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the current user
      tags:
      - users
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      summary: Register a user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ValidationError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIErrorResponse'
      summary: Log in
      tags:
      - users
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter `Bearer {access token}`.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"